requirements.txt
opencv-python
numpy
pygame
pyyaml
